WIP azure ai agents service migration
