+++ src/workshop/requirements.txt
@@ -7,3 +7,4 @@ azure-ai-projects==1.0.0b5
 pandas>=2.2.3, <3.0.0
 pydantic==2.10.1
 pillow>=11.1.0, <12.0.0
+chainlit==1.3.2
